Start splitting out carmen-cache functions into separate files; coalesce is done now and the util files are started
