+++ src/binding.cpp
@@ -1,5 +1,7 @@
 
 #include "binding.hpp"
+#include "node_util.hpp"
+#include "coalesce.hpp"
 
 #include <algorithm>
 #include <cassert>
@@ -35,8 +37,6 @@ rocksdb::Status OpenForReadOnlyDB(const rocksdb::Options& options, const std::st
     dbptr = std::unique_ptr<rocksdb::DB>(db);
     return status;
 }
-
-constexpr uint64_t LANGUAGE_MATCH_BOOST = (const uint64_t)(1) << 63;
 
 // in general, the key format including language field is <key_text><separator character><8-byte langfield>
 // but as a size optimization for language-less indexes, we omit the langfield
@@ -262,33 +262,6 @@ intarray __getmatching(RocksDBCache const* c, std::string phrase, bool match_pre
     }
 
     return array;
-}
-
-constexpr unsigned MAX_LANG = (sizeof(langfield_type) * 8) - 1;
-inline langfield_type langarrayToLangfield(Local<v8::Array> const& array) {
-    size_t array_size = array->Length();
-    langfield_type out = 0;
-    for (unsigned i = 0; i < array_size; i++) {
-        unsigned int val = static_cast<unsigned int>(array->Get(i)->NumberValue());
-        if (val > MAX_LANG) {
-            // this should probably throw something
-            continue;
-        }
-        out = out | (static_cast<langfield_type>(1) << val);
-    }
-    return out;
-}
-
-inline Local<v8::Array> langfieldToLangarray(langfield_type langfield) {
-    Local<Array> langs = Nan::New<Array>();
-
-    unsigned idx = 0;
-    for (unsigned i = 0; i <= MAX_LANG; i++) {
-        if (langfield & (static_cast<langfield_type>(1) << i)) {
-            langs->Set(idx++, Nan::New(i));
-        }
-    }
-    return langs;
 }
 
 void MemoryCache::Initialize(Handle<Object> target) {
@@ -1006,959 +979,6 @@ NAN_METHOD(RocksDBCache::New) {
     } catch (std::exception const& ex) {
         return Nan::ThrowTypeError(ex.what());
     }
-    info.GetReturnValue().Set(Nan::Undefined());
-    return;
-}
-
-//relev = 5 bits
-//count = 3 bits
-//reason = 12 bits
-//* 1 bit gap
-//id = 32 bits
-constexpr double _pow(double x, int y) {
-    return y == 0 ? 1.0 : x * _pow(x, y - 1);
-}
-
-constexpr uint64_t POW2_51 = static_cast<uint64_t>(_pow(2.0, 51));
-constexpr uint64_t POW2_48 = static_cast<uint64_t>(_pow(2.0, 48));
-constexpr uint64_t POW2_34 = static_cast<uint64_t>(_pow(2.0, 34));
-constexpr uint64_t POW2_28 = static_cast<uint64_t>(_pow(2.0, 28));
-constexpr uint64_t POW2_25 = static_cast<uint64_t>(_pow(2.0, 25));
-constexpr uint64_t POW2_20 = static_cast<uint64_t>(_pow(2.0, 20));
-constexpr uint64_t POW2_14 = static_cast<uint64_t>(_pow(2.0, 14));
-constexpr uint64_t POW2_3 = static_cast<uint64_t>(_pow(2.0, 3));
-constexpr uint64_t POW2_2 = static_cast<uint64_t>(_pow(2.0, 2));
-
-struct PhrasematchSubq {
-    PhrasematchSubq(void* c,
-                    char t,
-                    double w,
-                    std::string p,
-                    bool pf,
-                    unsigned short i,
-                    unsigned short z,
-                    uint32_t m,
-                    langfield_type l) : cache(c),
-                                        type(t),
-                                        weight(w),
-                                        phrase(p),
-                                        prefix(pf),
-                                        idx(i),
-                                        zoom(z),
-                                        mask(m),
-                                        langfield(l) {}
-    void* cache;
-    char type;
-    double weight;
-    std::string phrase;
-    bool prefix;
-    unsigned short idx;
-    unsigned short zoom;
-    uint32_t mask;
-    langfield_type langfield;
-    PhrasematchSubq& operator=(PhrasematchSubq&& c) = default;
-    PhrasematchSubq(PhrasematchSubq&& c) = default;
-};
-
-struct Cover {
-    double relev;
-    uint32_t id;
-    uint32_t tmpid;
-    unsigned short x;
-    unsigned short y;
-    unsigned short score;
-    unsigned short idx;
-    uint32_t mask;
-    double distance;
-    double scoredist;
-    bool matches_language;
-};
-
-struct Context {
-    std::vector<Cover> coverList;
-    uint32_t mask;
-    double relev;
-
-    Context(Context const& c) = default;
-    Context(Cover&& cov,
-            uint32_t mask,
-            double relev)
-        : coverList(),
-          mask(mask),
-          relev(relev) {
-        coverList.emplace_back(std::move(cov));
-    }
-    Context& operator=(Context&& c) {
-        coverList = std::move(c.coverList);
-        mask = std::move(c.mask);
-        relev = std::move(c.relev);
-        return *this;
-    }
-    Context(std::vector<Cover>&& cl,
-            uint32_t mask,
-            double relev)
-        : coverList(std::move(cl)),
-          mask(mask),
-          relev(relev) {}
-
-    Context(Context&& c)
-        : coverList(std::move(c.coverList)),
-          mask(std::move(c.mask)),
-          relev(std::move(c.relev)) {}
-};
-
-Cover numToCover(uint64_t num) {
-    Cover cover;
-    assert(((num >> 34) % POW2_14) <= static_cast<double>(std::numeric_limits<unsigned short>::max()));
-    assert(((num >> 34) % POW2_14) >= static_cast<double>(std::numeric_limits<unsigned short>::min()));
-    unsigned short y = static_cast<unsigned short>((num >> 34) % POW2_14);
-    assert(((num >> 20) % POW2_14) <= static_cast<double>(std::numeric_limits<unsigned short>::max()));
-    assert(((num >> 20) % POW2_14) >= static_cast<double>(std::numeric_limits<unsigned short>::min()));
-    unsigned short x = static_cast<unsigned short>((num >> 20) % POW2_14);
-    assert(((num >> 48) % POW2_3) <= static_cast<double>(std::numeric_limits<unsigned short>::max()));
-    assert(((num >> 48) % POW2_3) >= static_cast<double>(std::numeric_limits<unsigned short>::min()));
-    unsigned short score = static_cast<unsigned short>((num >> 48) % POW2_3);
-    uint32_t id = static_cast<uint32_t>(num % POW2_20);
-    bool matches_language = static_cast<bool>(num & LANGUAGE_MATCH_BOOST);
-    cover.x = x;
-    cover.y = y;
-    double relev = 0.4 + (0.2 * static_cast<double>((num >> 51) % POW2_2));
-    cover.relev = relev;
-    cover.score = score;
-    cover.id = id;
-    cover.matches_language = matches_language;
-
-    // These are not derived from decoding the input num but by
-    // external values after initialization.
-    cover.idx = 0;
-    cover.mask = 0;
-    cover.tmpid = 0;
-    cover.distance = 0;
-
-    return cover;
-}
-
-struct ZXY {
-    unsigned z;
-    unsigned x;
-    unsigned y;
-};
-
-ZXY pxy2zxy(unsigned z, unsigned x, unsigned y, unsigned target_z) {
-    ZXY zxy;
-    zxy.z = target_z;
-
-    // Interval between parent and target zoom level
-    signed zDist = static_cast<signed>(target_z) - static_cast<signed>(z);
-    signed zMult = zDist - 1;
-    if (zDist == 0) {
-        zxy.x = x;
-        zxy.y = y;
-        return zxy;
-    }
-
-    // Midpoint length @ z for a tile at parent zoom level
-    double pMid_d = static_cast<double>(std::pow(2, zDist) / 2);
-    assert(pMid_d <= static_cast<double>(std::numeric_limits<unsigned>::max()));
-    assert(pMid_d >= static_cast<double>(std::numeric_limits<unsigned>::min()));
-    signed pMid = static_cast<signed>(pMid_d);
-    zxy.x = static_cast<unsigned>((static_cast<signed>(x) * zMult) + pMid);
-    zxy.y = static_cast<unsigned>((static_cast<signed>(y) * zMult) + pMid);
-    return zxy;
-}
-
-ZXY bxy2zxy(unsigned z, unsigned x, unsigned y, unsigned target_z, bool max = false) {
-    ZXY zxy;
-    zxy.z = target_z;
-
-    // Interval between parent and target zoom level
-    signed zDist = static_cast<signed>(target_z) - static_cast<signed>(z);
-    if (zDist == 0) {
-        zxy.x = x;
-        zxy.y = y;
-        return zxy;
-    }
-
-    // zoom conversion multiplier
-    float mult = static_cast<float>(std::pow(2, zDist));
-
-    // zoom in min
-    if (zDist > 0 && !max) {
-        zxy.x = static_cast<unsigned>(static_cast<float>(x) * mult);
-        zxy.y = static_cast<unsigned>(static_cast<float>(y) * mult);
-        return zxy;
-    }
-    // zoom in max
-    else if (zDist > 0 && max) {
-        zxy.x = static_cast<unsigned>(static_cast<float>(x) * mult + (mult - 1));
-        zxy.y = static_cast<unsigned>(static_cast<float>(y) * mult + (mult - 1));
-        return zxy;
-    }
-    // zoom out
-    else {
-        unsigned mod = static_cast<unsigned>(std::pow(2, target_z));
-        unsigned xDiff = x % mod;
-        unsigned yDiff = y % mod;
-        unsigned newX = x - xDiff;
-        unsigned newY = y - yDiff;
-
-        zxy.x = static_cast<unsigned>(static_cast<float>(newX) * mult);
-        zxy.y = static_cast<unsigned>(static_cast<float>(newY) * mult);
-        return zxy;
-    }
-}
-
-inline bool coverSortByRelev(Cover const& a, Cover const& b) noexcept {
-    if (b.relev > a.relev)
-        return false;
-    else if (b.relev < a.relev)
-        return true;
-    else if (b.scoredist > a.scoredist)
-        return false;
-    else if (b.scoredist < a.scoredist)
-        return true;
-    else if (b.idx < a.idx)
-        return false;
-    else if (b.idx > a.idx)
-        return true;
-    else if (b.id < a.id)
-        return false;
-    else if (b.id > a.id)
-        return true;
-    // sorting by x and y is arbitrary but provides a more deterministic output order
-    else if (b.x < a.x)
-        return false;
-    else if (b.x > a.x)
-        return true;
-    else
-        return (b.y > a.y);
-}
-
-inline bool subqSortByZoom(PhrasematchSubq const& a, PhrasematchSubq const& b) noexcept {
-    if (a.zoom < b.zoom) return true;
-    if (a.zoom > b.zoom) return false;
-    return (a.idx < b.idx);
-}
-
-inline bool contextSortByRelev(Context const& a, Context const& b) noexcept {
-    if (b.relev > a.relev)
-        return false;
-    else if (b.relev < a.relev)
-        return true;
-    else if (b.coverList[0].scoredist > a.coverList[0].scoredist)
-        return false;
-    else if (b.coverList[0].scoredist < a.coverList[0].scoredist)
-        return true;
-    else if (b.coverList[0].idx < a.coverList[0].idx)
-        return false;
-    else if (b.coverList[0].idx > a.coverList[0].idx)
-        return true;
-    return (b.coverList[0].id > a.coverList[0].id);
-}
-
-inline double tileDist(unsigned px, unsigned py, unsigned tileX, unsigned tileY) {
-    const double dx = static_cast<double>(px) - static_cast<double>(tileX);
-    const double dy = static_cast<double>(py) - static_cast<double>(tileY);
-    const double distance = sqrt((dx * dx) + (dy * dy));
-
-    return distance;
-}
-
-struct CoalesceBaton : carmen::noncopyable {
-    uv_work_t request;
-    // params
-    std::vector<PhrasematchSubq> stack;
-    std::vector<uint64_t> centerzxy;
-    std::vector<uint64_t> bboxzxy;
-    double radius;
-    Nan::Persistent<v8::Function> callback;
-    // return
-    std::vector<Context> features;
-    // error
-    std::string error;
-};
-
-// Equivalent of scoredist() function in carmen
-// Combines score and distance into a single score that can be used for sorting.
-// Unlike carmen the effect is not scaled by zoom level as regardless of index
-// the score value at this stage is a 0-7 scalar (by comparison, in carmen, scores
-// for indexes with features covering lower zooms often have exponentially higher
-// scores - example: country@z9 vs poi@z14).
-double scoredist(unsigned zoom, double distance, double score, double radius) {
-    if (zoom < 6) zoom = 6;
-    if (distance == 0.0) distance = 0.01;
-    double scoredist = 0;
-
-    // Since distance is in tiles we calculate scoredist by converting the miles into
-    // a tile unit value at the appropriate zoom first.
-    //
-    // 32 tiles is about 40 miles at z14, use this as our mile <=> tile conversion.
-    scoredist = ((radius * (32.0 / 40.0)) / _pow(1.5, 14 - static_cast<int>(zoom))) / distance;
-
-    return score > scoredist ? score : scoredist;
-}
-
-void coalesceFinalize(CoalesceBaton* baton, std::vector<Context>&& contexts) {
-    if (!contexts.empty()) {
-        // Coalesce stack, generate relevs.
-        double relevMax = contexts[0].relev;
-        std::size_t total = 0;
-        std::map<uint64_t, bool> sets;
-        std::map<uint64_t, bool>::iterator sit;
-        std::size_t max_contexts = 40;
-        baton->features.reserve(max_contexts);
-        for (auto&& context : contexts) {
-            // Maximum allowance of coalesced features: 40.
-            if (total >= max_contexts) break;
-
-            // Since `coalesced` is sorted by relev desc at first
-            // threshold miss we can break the loop.
-            if (relevMax - context.relev >= 0.25) break;
-
-            // Only collect each feature once.
-            uint32_t id = context.coverList[0].tmpid;
-            sit = sets.find(id);
-            if (sit != sets.end()) continue;
-
-            sets.emplace(id, true);
-            baton->features.emplace_back(std::move(context));
-            total++;
-        }
-    }
-}
-void coalesceSingle(uv_work_t* req) {
-    CoalesceBaton* baton = static_cast<CoalesceBaton*>(req->data);
-
-    try {
-        std::vector<PhrasematchSubq> const& stack = baton->stack;
-        PhrasematchSubq const& subq = stack[0];
-
-        // proximity (optional)
-        bool proximity = !baton->centerzxy.empty();
-        unsigned cz;
-        unsigned cx;
-        unsigned cy;
-        if (proximity) {
-            cz = static_cast<unsigned>(baton->centerzxy[0]);
-            cx = static_cast<unsigned>(baton->centerzxy[1]);
-            cy = static_cast<unsigned>(baton->centerzxy[2]);
-        } else {
-            cz = 0;
-            cx = 0;
-            cy = 0;
-        }
-
-        // bbox (optional)
-        bool bbox = !baton->bboxzxy.empty();
-        unsigned minx;
-        unsigned miny;
-        unsigned maxx;
-        unsigned maxy;
-        if (bbox) {
-            minx = static_cast<unsigned>(baton->bboxzxy[1]);
-            miny = static_cast<unsigned>(baton->bboxzxy[2]);
-            maxx = static_cast<unsigned>(baton->bboxzxy[3]);
-            maxy = static_cast<unsigned>(baton->bboxzxy[4]);
-        } else {
-            minx = 0;
-            miny = 0;
-            maxx = 0;
-            maxy = 0;
-        }
-
-        // Load and concatenate grids for all ids in `phrases`
-        intarray grids;
-        grids = subq.type == TYPE_MEMORY ? __getmatching(reinterpret_cast<MemoryCache*>(subq.cache), subq.phrase, subq.prefix, subq.langfield) : __getmatching(reinterpret_cast<RocksDBCache*>(subq.cache), subq.phrase, subq.prefix, subq.langfield);
-
-        unsigned long m = grids.size();
-        double relevMax = 0;
-        std::vector<Cover> covers;
-        covers.reserve(m);
-
-        uint32_t length = 0;
-        uint32_t lastId = 0;
-        double lastRelev = 0;
-        double lastScoredist = 0;
-        double lastDistance = 0;
-        double minScoredist = std::numeric_limits<double>::max();
-        for (unsigned long j = 0; j < m; j++) {
-            Cover cover = numToCover(grids[j]);
-
-            cover.idx = subq.idx;
-            cover.tmpid = static_cast<uint32_t>(cover.idx * POW2_25 + cover.id);
-            cover.relev = cover.relev * subq.weight;
-            if (!cover.matches_language) cover.relev *= .9;
-            cover.distance = proximity ? tileDist(cx, cy, cover.x, cover.y) : 0;
-            cover.scoredist = proximity ? scoredist(cz, cover.distance, cover.score, baton->radius) : cover.score;
-
-            // only add cover id if it's got a higer scoredist
-            if (lastId == cover.id && cover.scoredist <= lastScoredist) continue;
-
-            // short circuit based on relevMax thres
-            if (length > 40) {
-                if (cover.scoredist < minScoredist) continue;
-                if (cover.relev < lastRelev) break;
-            }
-            if (relevMax - cover.relev >= 0.25) break;
-            if (cover.relev > relevMax) relevMax = cover.relev;
-
-            if (bbox) {
-                if (cover.x < minx || cover.y < miny || cover.x > maxx || cover.y > maxy) continue;
-            }
-
-            covers.emplace_back(cover);
-            if (lastId != cover.id) length++;
-            if (!proximity && length > 40) break;
-            if (cover.scoredist < minScoredist) minScoredist = cover.scoredist;
-            lastId = cover.id;
-            lastRelev = cover.relev;
-            lastScoredist = cover.scoredist;
-            lastDistance = cover.distance;
-        }
-
-        // sort grids by distance to proximity point
-        std::sort(covers.begin(), covers.end(), coverSortByRelev);
-
-        uint32_t lastid = 0;
-        std::size_t added = 0;
-        std::vector<Context> contexts;
-        std::size_t max_contexts = 40;
-        contexts.reserve(max_contexts);
-        for (auto&& cover : covers) {
-            // Stop at 40 contexts
-            if (added == max_contexts) break;
-
-            // Attempt not to add the same feature but by diff cover twice
-            if (lastid == cover.id) continue;
-
-            lastid = cover.id;
-            added++;
-
-            double relev = cover.relev;
-            uint32_t mask = 0;
-            contexts.emplace_back(std::move(cover), mask, relev);
-        }
-
-        coalesceFinalize(baton, std::move(contexts));
-    } catch (std::exception const& ex) {
-        baton->error = ex.what();
-    }
-}
-
-void coalesceMulti(uv_work_t* req) {
-    CoalesceBaton* baton = static_cast<CoalesceBaton*>(req->data);
-
-    try {
-        std::vector<PhrasematchSubq>& stack = baton->stack;
-        std::sort(stack.begin(), stack.end(), subqSortByZoom);
-        std::size_t stackSize = stack.size();
-
-        // Cache zoom levels to iterate over as coalesce occurs.
-        std::vector<intarray> zoomCache;
-        zoomCache.reserve(stackSize);
-        for (auto const& subq : stack) {
-            intarray zooms;
-            std::vector<bool> zoomUniq(22, false);
-            for (auto const& subqB : stack) {
-                if (subq.idx == subqB.idx) continue;
-                if (zoomUniq[subqB.zoom]) continue;
-                if (subq.zoom < subqB.zoom) continue;
-                zoomUniq[subqB.zoom] = true;
-                zooms.emplace_back(subqB.zoom);
-            }
-            zoomCache.push_back(std::move(zooms));
-        }
-
-        // Coalesce relevs into higher zooms, e.g.
-        // z5 inherits relev of overlapping tiles at z4.
-        // @TODO assumes sources are in zoom ascending order.
-        std::map<uint64_t, std::vector<Context>> coalesced;
-        std::map<uint64_t, std::vector<Context>>::iterator cit;
-        std::map<uint64_t, std::vector<Context>>::iterator pit;
-        std::map<uint64_t, bool> done;
-        std::map<uint64_t, bool>::iterator dit;
-
-        // proximity (optional)
-        bool proximity = baton->centerzxy.size() > 0;
-        unsigned cz;
-        unsigned cx;
-        unsigned cy;
-        if (proximity) {
-            cz = static_cast<unsigned>(baton->centerzxy[0]);
-            cx = static_cast<unsigned>(baton->centerzxy[1]);
-            cy = static_cast<unsigned>(baton->centerzxy[2]);
-        } else {
-            cz = 0;
-            cx = 0;
-            cy = 0;
-        }
-
-        // bbox (optional)
-        bool bbox = !baton->bboxzxy.empty();
-        unsigned bboxz;
-        unsigned minx;
-        unsigned miny;
-        unsigned maxx;
-        unsigned maxy;
-        if (bbox) {
-            bboxz = static_cast<unsigned>(baton->bboxzxy[0]);
-            minx = static_cast<unsigned>(baton->bboxzxy[1]);
-            miny = static_cast<unsigned>(baton->bboxzxy[2]);
-            maxx = static_cast<unsigned>(baton->bboxzxy[3]);
-            maxy = static_cast<unsigned>(baton->bboxzxy[4]);
-        } else {
-            bboxz = 0;
-            minx = 0;
-            miny = 0;
-            maxx = 0;
-            maxy = 0;
-        }
-
-        std::vector<Context> contexts;
-        std::size_t i = 0;
-        for (auto const& subq : stack) {
-            // Load and concatenate grids for all ids in `phrases`
-            intarray grids;
-            grids = subq.type == TYPE_MEMORY ? __getmatching(reinterpret_cast<MemoryCache*>(subq.cache), subq.phrase, subq.prefix, subq.langfield) : __getmatching(reinterpret_cast<RocksDBCache*>(subq.cache), subq.phrase, subq.prefix, subq.langfield);
-
-            bool first = i == 0;
-            bool last = i == (stack.size() - 1);
-            unsigned short z = subq.zoom;
-            auto const& zCache = zoomCache[i];
-            std::size_t zCacheSize = zCache.size();
-
-            unsigned long m = grids.size();
-
-            for (unsigned long j = 0; j < m; j++) {
-                Cover cover = numToCover(grids[j]);
-                cover.idx = subq.idx;
-                cover.mask = subq.mask;
-                cover.tmpid = static_cast<uint32_t>(cover.idx * POW2_25 + cover.id);
-                cover.relev = cover.relev * subq.weight;
-                if (!cover.matches_language) cover.relev *= .9;
-                if (proximity) {
-                    ZXY dxy = pxy2zxy(z, cover.x, cover.y, cz);
-                    cover.distance = tileDist(cx, cy, dxy.x, dxy.y);
-                    cover.scoredist = scoredist(cz, cover.distance, cover.score, baton->radius);
-                } else {
-                    cover.distance = 0;
-                    cover.scoredist = cover.score;
-                }
-
-                if (bbox) {
-                    ZXY min = bxy2zxy(bboxz, minx, miny, z, false);
-                    ZXY max = bxy2zxy(bboxz, maxx, maxy, z, true);
-                    if (cover.x < min.x || cover.y < min.y || cover.x > max.x || cover.y > max.y) continue;
-                }
-
-                uint64_t zxy = (z * POW2_28) + (cover.x * POW2_14) + (cover.y);
-
-                // Reserve stackSize for the coverList. The vector
-                // will grow no larger that the size of the input
-                // subqueries that are being coalesced.
-                std::vector<Cover> covers;
-                covers.reserve(stackSize);
-                covers.push_back(cover);
-                uint32_t context_mask = cover.mask;
-                double context_relev = cover.relev;
-
-                for (unsigned a = 0; a < zCacheSize; a++) {
-                    uint64_t p = zCache[a];
-                    double s = static_cast<double>(1 << (z - p));
-                    uint64_t pxy = static_cast<uint64_t>(p * POW2_28) +
-                                   static_cast<uint64_t>(std::floor(cover.x / s) * POW2_14) +
-                                   static_cast<uint64_t>(std::floor(cover.y / s));
-                    pit = coalesced.find(pxy);
-                    if (pit != coalesced.end()) {
-                        uint32_t lastMask = 0;
-                        double lastRelev = 0.0;
-                        for (auto const& parents : pit->second) {
-                            for (auto const& parent : parents.coverList) {
-                                // this cover is functionally identical with previous and
-                                // is more relevant, replace the previous.
-                                if (parent.mask == lastMask && parent.relev > lastRelev) {
-                                    covers.pop_back();
-                                    covers.emplace_back(parent);
-                                    context_relev -= lastRelev;
-                                    context_relev += parent.relev;
-                                    lastMask = parent.mask;
-                                    lastRelev = parent.relev;
-                                    // this cover doesn't overlap with used mask.
-                                } else if (!(context_mask & parent.mask)) {
-                                    covers.emplace_back(parent);
-                                    context_relev += parent.relev;
-                                    context_mask = context_mask | parent.mask;
-                                    lastMask = parent.mask;
-                                    lastRelev = parent.relev;
-                                }
-                            }
-                        }
-                    }
-                }
-
-                if (last) {
-                    // Slightly penalize contexts that have no stacking
-                    if (covers.size() == 1) {
-                        context_relev -= 0.01;
-                        // Slightly penalize contexts in ascending order
-                    } else if (covers[0].mask > covers[1].mask) {
-                        context_relev -= 0.01;
-                    }
-                    contexts.emplace_back(std::move(covers), context_mask, context_relev);
-                } else if (first || covers.size() > 1) {
-                    cit = coalesced.find(zxy);
-                    if (cit == coalesced.end()) {
-                        std::vector<Context> local_contexts;
-                        local_contexts.emplace_back(std::move(covers), context_mask, context_relev);
-                        coalesced.emplace(zxy, std::move(local_contexts));
-                    } else {
-                        cit->second.emplace_back(std::move(covers), context_mask, context_relev);
-                    }
-                }
-            }
-
-            i++;
-        }
-
-        // append coalesced to contexts by moving memory
-        for (auto&& matched : coalesced) {
-            for (auto&& context : matched.second) {
-                contexts.emplace_back(std::move(context));
-            }
-        }
-
-        std::sort(contexts.begin(), contexts.end(), contextSortByRelev);
-        coalesceFinalize(baton, std::move(contexts));
-    } catch (std::exception const& ex) {
-        baton->error = ex.what();
-    }
-}
-
-Local<Object> coverToObject(Cover const& cover) {
-    Local<Object> object = Nan::New<Object>();
-    object->Set(Nan::New("x").ToLocalChecked(), Nan::New<Number>(cover.x));
-    object->Set(Nan::New("y").ToLocalChecked(), Nan::New<Number>(cover.y));
-    object->Set(Nan::New("relev").ToLocalChecked(), Nan::New<Number>(cover.relev));
-    object->Set(Nan::New("score").ToLocalChecked(), Nan::New<Number>(cover.score));
-    object->Set(Nan::New("id").ToLocalChecked(), Nan::New<Number>(cover.id));
-    object->Set(Nan::New("idx").ToLocalChecked(), Nan::New<Number>(cover.idx));
-    object->Set(Nan::New("tmpid").ToLocalChecked(), Nan::New<Number>(cover.tmpid));
-    object->Set(Nan::New("distance").ToLocalChecked(), Nan::New<Number>(cover.distance));
-    object->Set(Nan::New("scoredist").ToLocalChecked(), Nan::New<Number>(cover.scoredist));
-    object->Set(Nan::New("matches_language").ToLocalChecked(), Nan::New<Boolean>(cover.matches_language));
-    return object;
-}
-Local<Array> contextToArray(Context const& context) {
-    std::size_t size = context.coverList.size();
-    Local<Array> array = Nan::New<Array>(static_cast<int>(size));
-    for (uint32_t i = 0; i < size; i++) {
-        array->Set(i, coverToObject(context.coverList[i]));
-    }
-    array->Set(Nan::New("relev").ToLocalChecked(), Nan::New(context.relev));
-    return array;
-}
-void coalesceAfter(uv_work_t* req) {
-    Nan::HandleScope scope;
-    CoalesceBaton* baton = static_cast<CoalesceBaton*>(req->data);
-
-    // Reference count the cache objects
-    for (auto& subq : baton->stack) {
-        if (subq.type == TYPE_MEMORY)
-            reinterpret_cast<MemoryCache*>(subq.cache)->_unref();
-        else
-            reinterpret_cast<RocksDBCache*>(subq.cache)->_unref();
-    }
-
-    if (!baton->error.empty()) {
-        v8::Local<v8::Value> argv[1] = {Nan::Error(baton->error.c_str())};
-        Nan::MakeCallback(Nan::GetCurrentContext()->Global(), Nan::New(baton->callback), 1, argv);
-    } else {
-        std::vector<Context> const& features = baton->features;
-
-        Local<Array> jsFeatures = Nan::New<Array>(static_cast<int>(features.size()));
-        for (std::size_t i = 0; i < features.size(); i++) {
-            jsFeatures->Set(i, contextToArray(features[i]));
-        }
-
-        Local<Value> argv[2] = {Nan::Null(), jsFeatures};
-        Nan::MakeCallback(Nan::GetCurrentContext()->Global(), Nan::New(baton->callback), 2, argv);
-    }
-
-    baton->callback.Reset();
-    delete baton;
-}
-
-NAN_METHOD(coalesce) {
-    // PhrasematchStack (js => cpp)
-    if (info.Length() < 3) {
-        return Nan::ThrowTypeError("Expects 3 arguments: a PhrasematchSubq array, an option object, and a callback");
-    }
-
-    if (!info[0]->IsArray()) {
-        return Nan::ThrowTypeError("Arg 1 must be a PhrasematchSubq array");
-    }
-
-    Local<Array> array = Local<Array>::Cast(info[0]);
-    auto array_length = array->Length();
-    if (array_length < 1) {
-        return Nan::ThrowTypeError("Arg 1 must be an array with one or more PhrasematchSubq objects");
-    }
-
-    // Options object (js => cpp)
-    Local<Value> options_val = info[1];
-    if (!options_val->IsObject()) {
-        return Nan::ThrowTypeError("Arg 2 must be an options object");
-    }
-    Local<Object> options = options_val->ToObject();
-
-    // callback
-    Local<Value> callback = info[2];
-    if (!callback->IsFunction()) {
-        return Nan::ThrowTypeError("Arg 3 must be a callback function");
-    }
-
-    // We use unique_ptr here to manage the heap allocated CoalesceBaton
-    // If an error is thrown the unique_ptr will go out of scope and delete
-    // its underlying baton.
-    // If no error is throw we release the underlying baton pointer before
-    // heading into the threadpool since we assume it will be deleted manually in coalesceAfter
-    std::unique_ptr<CoalesceBaton> baton_ptr = std::make_unique<CoalesceBaton>();
-    CoalesceBaton* baton = baton_ptr.get();
-    try {
-        for (uint32_t i = 0; i < array_length; i++) {
-            Local<Value> val = array->Get(i);
-            if (!val->IsObject()) {
-                return Nan::ThrowTypeError("All items in array must be valid PhrasematchSubq objects");
-            }
-            Local<Object> jsStack = val->ToObject();
-            if (jsStack->IsNull() || jsStack->IsUndefined()) {
-                return Nan::ThrowTypeError("All items in array must be valid PhrasematchSubq objects");
-            }
-
-            double weight;
-            std::string phrase;
-            bool prefix;
-            unsigned short idx;
-            unsigned short zoom;
-            uint32_t mask;
-            langfield_type langfield;
-
-            // TODO: this is verbose: we could write some generic functions to do this robust conversion per type
-            if (!jsStack->Has(Nan::New("idx").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing idx property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("idx").ToLocalChecked());
-                if (!prop_val->IsNumber()) {
-                    return Nan::ThrowTypeError("idx value must be a number");
-                }
-                int64_t _idx = prop_val->IntegerValue();
-                if (_idx < 0 || _idx > std::numeric_limits<unsigned short>::max()) {
-                    return Nan::ThrowTypeError("encountered idx value too large to fit in unsigned short");
-                }
-                idx = static_cast<unsigned short>(_idx);
-            }
-
-            if (!jsStack->Has(Nan::New("zoom").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing zoom property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("zoom").ToLocalChecked());
-                if (!prop_val->IsNumber()) {
-                    return Nan::ThrowTypeError("zoom value must be a number");
-                }
-                int64_t _zoom = prop_val->IntegerValue();
-                if (_zoom < 0 || _zoom > std::numeric_limits<unsigned short>::max()) {
-                    return Nan::ThrowTypeError("encountered zoom value too large to fit in unsigned short");
-                }
-                zoom = static_cast<unsigned short>(_zoom);
-            }
-
-            if (!jsStack->Has(Nan::New("weight").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing weight property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("weight").ToLocalChecked());
-                if (!prop_val->IsNumber()) {
-                    return Nan::ThrowTypeError("weight value must be a number");
-                }
-                double _weight = prop_val->NumberValue();
-                if (_weight < 0 || _weight > std::numeric_limits<double>::max()) {
-                    return Nan::ThrowTypeError("encountered weight value too large to fit in double");
-                }
-                weight = _weight;
-            }
-
-            if (!jsStack->Has(Nan::New("phrase").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing phrase property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("phrase").ToLocalChecked());
-                if (!prop_val->IsString()) {
-                    return Nan::ThrowTypeError("phrase value must be a string");
-                }
-                Nan::Utf8String _phrase(prop_val);
-                if (_phrase.length() < 1) {
-                    return Nan::ThrowTypeError("encountered invalid phrase");
-                }
-                phrase = *_phrase;
-            }
-
-            if (!jsStack->Has(Nan::New("prefix").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing prefix property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("prefix").ToLocalChecked());
-                if (!prop_val->IsBoolean()) {
-                    return Nan::ThrowTypeError("prefix value must be a boolean");
-                }
-                prefix = prop_val->BooleanValue();
-            }
-
-            if (!jsStack->Has(Nan::New("mask").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing mask property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("mask").ToLocalChecked());
-                if (!prop_val->IsNumber()) {
-                    return Nan::ThrowTypeError("mask value must be a number");
-                }
-                int64_t _mask = prop_val->IntegerValue();
-                if (_mask < 0 || _mask > std::numeric_limits<uint32_t>::max()) {
-                    return Nan::ThrowTypeError("encountered mask value too large to fit in uint32_t");
-                }
-                mask = static_cast<uint32_t>(_mask);
-            }
-
-            langfield = ALL_LANGUAGES;
-            if (jsStack->Has(Nan::New("languages").ToLocalChecked())) {
-                Local<Value> c_array = jsStack->Get(Nan::New("languages").ToLocalChecked());
-                if (!c_array->IsArray()) {
-                    return Nan::ThrowTypeError("languages must be an array");
-                }
-                Local<Array> carray = Local<Array>::Cast(c_array);
-                langfield = langarrayToLangfield(carray);
-            }
-
-            if (!jsStack->Has(Nan::New("cache").ToLocalChecked())) {
-                return Nan::ThrowTypeError("missing cache property");
-            } else {
-                Local<Value> prop_val = jsStack->Get(Nan::New("cache").ToLocalChecked());
-                if (!prop_val->IsObject()) {
-                    return Nan::ThrowTypeError("cache value must be a Cache object");
-                }
-                Local<Object> _cache = prop_val->ToObject();
-                if (_cache->IsNull() || _cache->IsUndefined()) {
-                    return Nan::ThrowTypeError("cache value must be a Cache object");
-                }
-                bool isMemoryCache = Nan::New(MemoryCache::constructor)->HasInstance(prop_val);
-                bool isRocksDBCache = Nan::New(RocksDBCache::constructor)->HasInstance(prop_val);
-                if (!(isMemoryCache || isRocksDBCache)) {
-                    return Nan::ThrowTypeError("cache value must be a MemoryCache or RocksDBCache object");
-                }
-                if (isMemoryCache) {
-                    baton->stack.emplace_back(
-                        (void*)node::ObjectWrap::Unwrap<MemoryCache>(_cache),
-                        TYPE_MEMORY,
-                        weight,
-                        phrase,
-                        prefix,
-                        idx,
-                        zoom,
-                        mask,
-                        langfield);
-                } else {
-                    baton->stack.emplace_back(
-                        (void*)node::ObjectWrap::Unwrap<RocksDBCache>(_cache),
-                        TYPE_ROCKSDB,
-                        weight,
-                        phrase,
-                        prefix,
-                        idx,
-                        zoom,
-                        mask,
-                        langfield);
-                }
-            }
-        }
-
-        if (options->Has(Nan::New("radius").ToLocalChecked())) {
-            Local<Value> prop_val = options->Get(Nan::New("radius").ToLocalChecked());
-            if (!prop_val->IsNumber()) {
-                return Nan::ThrowTypeError("radius must be a number");
-            }
-            int64_t _radius = prop_val->IntegerValue();
-            if (_radius < 0 || _radius > std::numeric_limits<unsigned>::max()) {
-                return Nan::ThrowTypeError("encountered radius too large to fit in unsigned");
-            }
-            baton->radius = static_cast<double>(_radius);
-        } else {
-            baton->radius = 40.0;
-        }
-
-        if (options->Has(Nan::New("centerzxy").ToLocalChecked())) {
-            Local<Value> c_array = options->Get(Nan::New("centerzxy").ToLocalChecked());
-            if (!c_array->IsArray()) {
-                return Nan::ThrowTypeError("centerzxy must be an array");
-            }
-            Local<Array> carray = Local<Array>::Cast(c_array);
-            if (carray->Length() != 3) {
-                return Nan::ThrowTypeError("centerzxy must be an array of 3 numbers");
-            }
-            baton->centerzxy.reserve(carray->Length());
-            for (uint32_t i = 0; i < carray->Length(); ++i) {
-                Local<Value> item = carray->Get(i);
-                if (!item->IsNumber()) {
-                    return Nan::ThrowTypeError("centerzxy values must be number");
-                }
-                int64_t a_val = item->IntegerValue();
-                if (a_val < 0 || a_val > std::numeric_limits<uint32_t>::max()) {
-                    return Nan::ThrowTypeError("encountered centerzxy value too large to fit in uint32_t");
-                }
-                baton->centerzxy.emplace_back(static_cast<uint32_t>(a_val));
-            }
-        }
-
-        if (options->Has(Nan::New("bboxzxy").ToLocalChecked())) {
-            Local<Value> c_array = options->Get(Nan::New("bboxzxy").ToLocalChecked());
-            if (!c_array->IsArray()) {
-                return Nan::ThrowTypeError("bboxzxy must be an array");
-            }
-            Local<Array> carray = Local<Array>::Cast(c_array);
-            if (carray->Length() != 5) {
-                return Nan::ThrowTypeError("bboxzxy must be an array of 5 numbers");
-            }
-            baton->bboxzxy.reserve(carray->Length());
-            for (uint32_t i = 0; i < carray->Length(); ++i) {
-                Local<Value> item = carray->Get(i);
-                if (!item->IsNumber()) {
-                    return Nan::ThrowTypeError("bboxzxy values must be number");
-                }
-                int64_t a_val = item->IntegerValue();
-                if (a_val < 0 || a_val > std::numeric_limits<uint32_t>::max()) {
-                    return Nan::ThrowTypeError("encountered bboxzxy value too large to fit in uint32_t");
-                }
-                baton->bboxzxy.emplace_back(static_cast<uint32_t>(a_val));
-            }
-        }
-
-        baton->callback.Reset(callback.As<Function>());
-
-        // queue work
-        baton->request.data = baton;
-        // Release the managed baton
-        baton_ptr.release();
-        // Reference count the cache objects
-        for (auto& subq : baton->stack) {
-            if (subq.type == TYPE_MEMORY)
-                reinterpret_cast<MemoryCache*>(subq.cache)->_ref();
-            else
-                reinterpret_cast<RocksDBCache*>(subq.cache)->_ref();
-        }
-        // optimization: for stacks of 1, use coalesceSingle
-        if (baton->stack.size() == 1) {
-            uv_queue_work(uv_default_loop(), &baton->request, coalesceSingle, (uv_after_work_cb)coalesceAfter);
-        } else {
-            uv_queue_work(uv_default_loop(), &baton->request, coalesceMulti, (uv_after_work_cb)coalesceAfter);
-        }
-    } catch (std::exception const& ex) {
-        return Nan::ThrowTypeError(ex.what());
-    }
-
     info.GetReturnValue().Set(Nan::Undefined());
     return;
 }
