+++ src/node_util.cpp
@@ -1,0 +1,32 @@
+
+#include "node_util.hpp"
+
+namespace carmen {
+
+using namespace v8;
+
+Local<Object> coverToObject(Cover const& cover) {
+    Local<Object> object = Nan::New<Object>();
+    object->Set(Nan::New("x").ToLocalChecked(), Nan::New<Number>(cover.x));
+    object->Set(Nan::New("y").ToLocalChecked(), Nan::New<Number>(cover.y));
+    object->Set(Nan::New("relev").ToLocalChecked(), Nan::New<Number>(cover.relev));
+    object->Set(Nan::New("score").ToLocalChecked(), Nan::New<Number>(cover.score));
+    object->Set(Nan::New("id").ToLocalChecked(), Nan::New<Number>(cover.id));
+    object->Set(Nan::New("idx").ToLocalChecked(), Nan::New<Number>(cover.idx));
+    object->Set(Nan::New("tmpid").ToLocalChecked(), Nan::New<Number>(cover.tmpid));
+    object->Set(Nan::New("distance").ToLocalChecked(), Nan::New<Number>(cover.distance));
+    object->Set(Nan::New("scoredist").ToLocalChecked(), Nan::New<Number>(cover.scoredist));
+    object->Set(Nan::New("matches_language").ToLocalChecked(), Nan::New<Boolean>(cover.matches_language));
+    return object;
+}
+Local<Array> contextToArray(Context const& context) {
+    std::size_t size = context.coverList.size();
+    Local<Array> array = Nan::New<Array>(static_cast<int>(size));
+    for (uint32_t i = 0; i < size; i++) {
+        array->Set(i, coverToObject(context.coverList[i]));
+    }
+    array->Set(Nan::New("relev").ToLocalChecked(), Nan::New(context.relev));
+    return array;
+}
+
+} // namespace carmen
